维护工具

+++ bio-drqi-bsm/bio-drqi-bsm-dto/src/main/java/com/bio/drqi/bsm/req/BmsProductStockInLogReturnStockReqDTO.java
@@ -2,12 +2,14 @@ package com.bio.drqi.bsm.req;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BmsProductStockInLogReturnStockReqDTO {
 
     private Integer  id;
 
-    private Integer returnNumber;
+    private BigDecimal returnNumber;
 
     private String remark;
 }
